V1.2 统计表单数据接口

+++ marketing-tool-service/src/main/java/cn/rongcapital/mkt/service/ContactKeyListGetService.java
@@ -7,6 +7,6 @@ import cn.rongcapital.mkt.vo.out.GetContactKeyListOutput;
  */
 public interface ContactKeyListGetService {
 
-    GetContactKeyListOutput getContactKeyList(Integer contactId);
+    GetContactKeyListOutput getContactKeyList(Long contactId);
 
 }
